requirements.txt
google-generativeai
chromadb
pypdf
langchain-google-genai
python-dotenv
streamlit
langchain
langchain-community
pypdf2